+++ plugins/auth-backend-module-pinniped-provider/src/module.test.ts
@@ -15,7 +15,7 @@
  */
 import {
   mockServices,
-  setupRequestMockHandlers,
+  registerMswTestHooks,
   startTestBackend,
 } from '@backstage/backend-test-utils';
 import { Server } from 'http';
@@ -32,7 +32,7 @@ describe('authModulePinnipedProvider', () => {
   let publicKey: JWK;
 
   const mswServer = setupServer();
-  setupRequestMockHandlers(mswServer);
+  registerMswTestHooks(mswServer);
 
   const issuerMetadata = {
     issuer: 'https://pinniped.test',