+++ plugins/auth-node/src/passport/PassportHelpers.ts
@@ -40,6 +40,9 @@ export class PassportHelpers {
     if (profile.emails && profile.emails.length > 0) {
       const [firstEmail] = profile.emails;
       email = firstEmail.value;
+    } else if (profile.email) {
+      // This is the case for Atlassian
+      email = profile.email;
     }
 
     let picture: string | undefined = undefined;
@@ -48,6 +51,9 @@ export class PassportHelpers {
     } else if (profile.photos && profile.photos.length > 0) {
       const [firstPhoto] = profile.photos;
       picture = firstPhoto.value;
+    } else if (profile.photo) {
+      // This is the case for Atlassian
+      picture = profile.photo;
     }
 
     let displayName: string | undefined =