+++ plugins/kubernetes-backend/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @backstage/plugin-kubernetes-backend
 
+## 0.18.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.7.0
+  - @backstage/backend-common@0.23.3
+  - @backstage/plugin-permission-common@0.8.0
+  - @backstage/plugin-permission-node@0.8.0
+  - @backstage/plugin-auth-node@0.4.17
+  - @backstage/plugin-catalog-node@1.12.4
+  - @backstage/plugin-kubernetes-node@0.1.16
+  - @backstage/plugin-kubernetes-common@0.8.1
+  - @backstage/catalog-client@1.6.5
+  - @backstage/catalog-model@1.5.0
+  - @backstage/config@1.2.0
+  - @backstage/errors@1.2.4
+  - @backstage/integration-aws-node@0.1.12
+  - @backstage/types@1.1.1
+
 ## 0.18.3-next.1
 
 ### Patch Changes