+++ plugins/techdocs-backend/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @backstage/plugin-techdocs-backend
 
+## 1.10.9
+
+### Patch Changes
+
+- 9ecf5fd: Adds extension point for publishers to the techdocs backend
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.7.0
+  - @backstage/backend-common@0.23.3
+  - @backstage/plugin-permission-common@0.8.0
+  - @backstage/plugin-techdocs-node@1.12.8
+  - @backstage/integration@1.13.0
+  - @backstage/plugin-search-backend-module-techdocs@0.1.27
+  - @backstage/plugin-catalog-common@1.0.25
+  - @backstage/catalog-client@1.6.5
+  - @backstage/catalog-model@1.5.0
+  - @backstage/config@1.2.0
+  - @backstage/errors@1.2.4
+
 ## 1.10.9-next.1
 
 ### Patch Changes